Develop

+++ ppcls/arch/backbone/model_zoo/distilled_vision_transformer.py
@@ -89,8 +89,8 @@ class DistilledVisionTransformer(VisionTransformer):
         B = paddle.shape(x)[0]
         x = self.patch_embed(x)
 
-        cls_tokens = self.cls_token.expand((B, -1, -1))
-        dist_token = self.dist_token.expand((B, -1, -1))
+        cls_tokens = self.cls_token.expand((B, -1, -1)).astype(x.dtype)
+        dist_token = self.dist_token.expand((B, -1, -1)).astype(x.dtype)
         x = paddle.concat((cls_tokens, dist_token, x), axis=1)
 
         x = x + self.pos_embed
